refactor incomplete, new worktree

+++ src/dna/bitwise_utils.rs
@@ -1,3 +1,3 @@
-pub fn is_bit_on(activity_mask: u32, pos: usize) -> bool {
+pub fn is_bit_on(activity_mask: u128, pos: usize) -> bool {
     ((activity_mask >> pos) & 1) == 1
 }
